refactor(speech-corpus): move prebuilt corpus pipeline to workflows/prebuilt_speech_index
- Move speech_merge.py, speech_enrichment.py, build_speech_corpus.py and their
  tests into a new workflows/prebuilt_speech_index package (merge.py,
  enrichment.py, build.py)
- Move dedicated tests from tests/api_swedeb/core/ and
  tests/api_swedeb/workflows/ into tests/api_swedeb/workflows/prebuilt_speech_index/
- Update all import paths to use canonical new location (no shim files retained)
- Update CLI script (build_speech_corpus_cli.py) to canonical import path
- Isolates build-time pipeline from runtime core/ layer, preparing for a
  future move of the pipeline to the pyriksprot library
- All 79 tests pass
