+++ tests/api_swedeb/workflows/prebuilt_speech_index/test_build.py
@@ -11,7 +11,7 @@ import pandas as pd
 import pyarrow.feather as feather
 import pytest
 
-from api_swedeb.workflows.build_speech_corpus import SpeechCorpusBuilder, _iter_zip_paths, _load_zip, _process_zip
+from api_swedeb.workflows.prebuilt_speech_index.build import SpeechCorpusBuilder, _iter_zip_paths, _load_zip, _process_zip
 
 
 # ---------------------------------------------------------------------------
